Version 0.0.4.8: - Diverse kleinere Codekorrekturen basierend auf cppcheck-Ergebnissen: --  Debug-Code in dhcp.c hinter Präprozessordirektiven versteckt -- Vorzeichen von Variablen und in printf-Formatstrings korrigiert
git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@1392 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/pci.c
@@ -191,7 +191,7 @@ void pci_scan()
                     if (PCIdev->irq != 255)
                     {
                         char path[50];
-                        char* end = path + snprintf(path, 50, "PrettyOS/PCI/%u/", counter);
+                        char* end = path + snprintf(path, 50, "PrettyOS/PCI/%u/", (uint32_t)counter);
 
                         strcpy(end, "Bus");
                         int64_t Bus = bus;
